backend/requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
pgvector==0.2.5

# AI/ML
openai==1.12.0
sentence-transformers==2.3.1
numpy==1.24.4

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
pandas==2.0.3

# Task Queue
celery==5.3.6
redis==5.0.1

# Security
cryptography==42.0.5

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0

# Development
django-debug-toolbar==4.3.0
